+++ nix/materialized/x86_64-linux/project/.plan.nix/eventuo11y-dsl.nix
@@ -11,7 +11,7 @@
     flags = {};
     package = {
       specVersion = "3.0";
-      identifier = { name = "eventuo11y-dsl"; version = "0.1.0.0"; };
+      identifier = { name = "eventuo11y-dsl"; version = "0.2.0.0"; };
       license = "Apache-2.0";
       copyright = "Copyright 2022 Shea Levy.";
       maintainer = "shea@shealevy.com";
@@ -35,7 +35,6 @@
         depends = [
           (hsPkgs."base" or (errorHandler.buildDepError "base"))
           (hsPkgs."template-haskell" or (errorHandler.buildDepError "template-haskell"))
-          (hsPkgs."th-compat" or (errorHandler.buildDepError "th-compat"))
           ];
         buildable = true;
         modules = [