+++ eventuo11y-batteries/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Revision history for eventuo11y-batteries
 
-## 0.2.1.0 -- 2022-10-23
+## 0.2.1.2 -- 2022-10-31
+
+Update for eventu11y-json 0.2
+
+## 0.2.1.1 -- 2022-10-23
 
 Update for eventu11y 0.5.0.0