+++ eventuo11y-dsl/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for eventuo11y-dsl
 
+## 0.2.0.0 -- 2022-10-31
+
+- Alias `AnyQuote` to `Q` in old `template-haskell` for direct `[t|Foo|]` syntax
+
 ## 0.1.0.0 -- 2022-10-23
 
 Initial release