+++ eventuo11y-json/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for eventuo11y-json
 
+## 0.2.0.0 -- 2022-10-31
+
+- Update for 'AnyQuote' change in 'eventuo11y-dsl' version '0.2'
+
 ## 0.1.0.0 -- 2022-10-23
 
 Initial release